backend/requirements.txt
asgiref==3.4.1
autopep8==1.6.0
backports.zoneinfo==0.2.1
dateparser==1.1.0
Django==3.2.7
django-cors-headers==3.9.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
humanize==3.13.1
maya==0.6.1
pendulum==2.1.2
Pillow==8.3.2
pycodestyle==2.8.0
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
regex==2021.11.10
six==1.16.0
snaptime==0.2.4
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
tzlocal==4.1
